requirements.txt
requests==2.25.1
DateTime==4.4
tqdm==4.64.0
python-dotenv==0.20.0